pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "item"
version = "0.1.0"
description = "Stupidly simple Google Tasks CLI"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "google-api-python-client>=2.0",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.0",
]

[project.scripts]
item = "item.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
